course 2. excel basics  data set - Indian Startup Funding
This dataset has funding information of the Indian startups from January 2015 to August 2017. It includes columns with the date funded, the city the startup is based out of, the names of the funders, and the amount invested (in USD).

For more information on the values of individual fields, check out the Column Metadata.